tests/wpt/web-platform-tests/css/css-variables/variable-definition-border-shorthand-serialize.html

<!DOCTYPE html>
<html>
<head>
    <title>Variables - border shorthand set, serialize border-color</title>

    <meta rel="author" title="Kevin Babbitt">
    <meta rel="author" title="Greg Whitworth">
    <link rel="author" title="Microsoft Corporation" href="http://microsoft.com" />
    <link rel="help" href="http://www.w3.org/TR/css-variables-1/#variables-in-shorthands">

    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="target" style="border: var(--borderwidth) solid black"></div>

    <script type="text/javascript">
        "use strict";

        test(function() {
            var target = document.getElementById("target");
            assert_equals(target.style.getPropertyValue("border-color"), "");
        }, "border-color should serialize to empty when border shorthand references a variable");
    </script>
</body>
</html>
