third_party/WebKit/LayoutTests/external/wpt/css/css-variables/variable-definition-border-shorthand-serialize-expected.txt

This is a testharness.js-based test.
FAIL border-color should serialize to empty when border shorthand references a variable assert_equals: expected "" but got "var(--borderwidth) solid black"
Harness: the test ran to completion.
